feat: create vpc subnets almost dynamically based on input parameter
Now in the vpc.yaml template we allow the creation of three subnets based on the input parameter NumberOfAZs.
The technique using conditions is a bit cumersome since we need to do it with funky boolean logic.
It could be worth looking into macros and lambda functions to try improve this setup